+++ base/strings/search.jl
@@ -621,4 +621,14 @@ See also: [`contains`](@ref).
 occursin(needle::Union{AbstractString,AbstractChar}, haystack::AbstractString) =
     _searchindex(haystack, needle, firstindex(haystack)) != 0
 
+"""
+    occursin(haystack)
+
+Create a function that checks whether its argument occurs in `haystack`, i.e.
+a function equivalent to `needle -> occursin(needle, haystack)`.
+
+The returned function is of type `Base.Fix2{typeof(occursin)}`.
+"""
+occursin(haystack) = Base.Fix2(occursin, haystack)
+
 in(::AbstractString, ::AbstractString) = error("use occursin(x, y) for string containment")